+++ bot/speech.py
@@ -1,18 +1,6 @@
-from string import Template
-
 from bot.color import get_color
-from bot.message import IncomingMessage
-
-SPEECH_HTML = """
-<div class="speech">
-    $avatar
-    <svg class="tail">
-        <path d="M 0,15 15,15 15,0 M 15,0 C 15,4 4,15 0,15" />
-    </svg>
-    <div class="messages">
-        $content
-    </div>
-</div>"""
+from bot.message import IncomingMessage, Sticker
+from bot.utils import full_path, fill_template
 
 
 class Speech:
@@ -34,12 +22,19 @@ class Speech:
 
     pfp: str
     messages: list[IncomingMessage]
+    visibility: str = "visible"
 
     def draw(self) -> str:
-        str_template = Template(SPEECH_HTML)
-
         content = ""
         for message in self.messages:
             content += message.draw()
 
-        return str_template.substitute(content=content, avatar=self.pfp)
+        if type(self.messages[-1]) is Sticker:
+            self.visibility = "hidden"
+
+        return fill_template(
+            full_path("files/speech.html"),
+            content=content,
+            avatar=self.pfp,
+            visibility=self.visibility,
+        )