+++ bot/utils.py
@@ -1,5 +1,6 @@
 import os
 import pathlib
+from string import Template
 
 ROOT_PATH = pathlib.Path.cwd()
 
@@ -11,3 +12,9 @@ def full_path(path: str) -> str:
 def open_file(path: str) -> str:
     file = open(full_path(path), "r", encoding="utf8")
     return file.read()
+
+
+def fill_template(template_path: str, /, **kws) -> str:
+    template_text = open_file(template_path)
+    template = Template(template_text)
+    return template.substitute(**kws)